switched to bao-temp

+++ DAL/Repository/PostRepository.cs
@@ -10,49 +10,67 @@ namespace DAL.Repository
     {
         private readonly PregnancyTrackingSystemContext _context;
 
-        public PostRepository( )
+        public PostRepository(PregnancyTrackingSystemContext context)
         {
-            _context = new PregnancyTrackingSystemContext();
+            _context = context;
         }
 
-        public List<Post> GetAll()
-        {
-            return _context.Posts.Include(p => p.User).Include(p => p.Comments).ToList();
-        }
-
-        public Post GetById(int id)
+        public List<Post> GetAllPosts()
         {
             return _context.Posts
                 .Include(p => p.User)
                 .Include(p => p.Comments)
-                .FirstOrDefault(p => p.Id == id);
-        }
-
-        public List<Post> GetByUserId(int userId)
-        {
-            return _context.Posts
-                .Include(p => p.User)
-                .Include(p => p.Comments)
-                .Where(p => p.UserId == userId)
+                    .ThenInclude(c => c.User)
+                .OrderByDescending(p => p.CreatedAt)
                 .ToList();
         }
 
-        public void Add(Post post)
+        public Post GetPostById(int id)
+        {
+            return _context.Posts
+                .Include(p => p.User)
+                .Include(p => p.Comments)
+                    .ThenInclude(c => c.User)
+                .FirstOrDefault(p => p.Id == id);
+        }
+
+        public void AddPost(Post post)
         {
             _context.Posts.Add(post);
             _context.SaveChanges();
         }
 
-        public void Update(Post post)
+        public void AddComment(Comment comment)
         {
-            _context.Posts.Update(post);
+            _context.Comments.Add(comment);
             _context.SaveChanges();
         }
 
-        public void Delete(Post post)
+        public void DeletePost(int postId)
         {
-            _context.Posts.Remove(post);
-            _context.SaveChanges();
+            var post = _context.Posts.Find(postId);
+            if (post != null)
+            {
+                _context.Posts.Remove(post);
+                _context.SaveChanges();
+            }
+        }
+
+        public void DeleteComment(int commentId)
+        {
+            var comment = _context.Comments.Find(commentId);
+            if (comment != null)
+            {
+                _context.Comments.Remove(comment);
+                _context.SaveChanges();
+            }
+        }
+
+        public Comment GetCommentById(int commentId)
+        {
+            return _context.Comments
+                .Include(c => c.User)
+                .FirstOrDefault(c => c.Id == commentId);
         }
     }
-} 
+}
